API Reference

EasyJobsBase.Job — Type

Job(def::Thunk; description="", username="")

Create a simple job.

Arguments

def: a Thunk that encloses the job definition.

name: give a short name to the job.

description::String="": describe what the job does in more detail.

username::String="": indicate who executes the job.

Examples

julia> using Thinkers

julia> a = Job(Thunk(sleep, 5); username="me", description="Sleep for 5 seconds");

julia> b = Job(Thunk(run, `pwd` & `ls`); username="me", description="Run some commands");

EasyJobsBase.:← — Method

←(y, x)

Chain two Jobs reversely.

EasyJobsBase.:→ — Method

→(x, y)

Chain two Jobs.

EasyJobsBase.chain! — Method

chain!(x::Job, y::Job, z::Job...)

Chain multiple Jobs one after another.

EasyJobsBase.countexecution — Method

countexecution(job::Job)

Count how many times a Job has been run.

EasyJobsBase.creationtimeof — Method

Return the creation time of a Job.

EasyJobsBase.descriptionof — Method

descriptionof(job::Job)

Return the description of a Job.

EasyJobsBase.endtimeof — Method

endtimeof(job::Job)

Return the end time of a Job. Return nothing if it has not exited.

EasyJobsBase.getstatus — Method

getstatus(x::Job)

Get the current status of a Job.

EasyJobsBase.isexited — Method

Test if the Job has exited.

EasyJobsBase.isfailed — Method

Test if the Job failed during running.

EasyJobsBase.isinterrupted — Method

Test if the Job was interrupted during running.

EasyJobsBase.ispending — Method

Test if the Job is still pending.

EasyJobsBase.isrunning — Method

Test if the Job is running.

EasyJobsBase.issucceeded — Method

Test if the Job was successfully run.

EasyJobsBase.kill! — Method

kill!(exe::Executor)

Manually kill a Job, works only if it is running.

EasyJobsBase.listexited — Method

listexited(jobs)

Filter only the exited jobs in a sequence of Jobs.

EasyJobsBase.listfailed — Method

listfailed(jobs)

Filter only the failed jobs in a sequence of Jobs.

EasyJobsBase.listinterrupted — Method

listinterrupted(jobs)

Filter only the interrupted jobs in a sequence of Jobs.

EasyJobsBase.listpending — Method

listpending(jobs)

Filter only the pending jobs in a sequence of Jobs.

EasyJobsBase.listrunning — Method

listrunning(jobs)

Filter only the running jobs in a sequence of Jobs.

EasyJobsBase.listsucceeded — Method

listsucceeded(jobs)

Filter only the succeeded jobs in a sequence of Jobs.

EasyJobsBase.run! — Method

run!(job::Job; maxattempts=1, interval=1, waitfor=0)

Run a Job with a maximum number of attempts, with each attempt separated by a few seconds.

EasyJobsBase.starttimeof — Method

starttimeof(job::Job)

Return the start time of a Job. Return nothing if it is still pending.

EasyJobsBase.timecostof — Method

timecostof(job::Job)

Return the time cost of a Job since it started running.

If nothing, the Job is still pending. If it is finished, return how long it took to complete.

Thinkers.getresult — Method

getresult(job::Job)

Get the running result of a Job.

The result is wrapped by a Some type. Use something to retrieve its value. If it is nothing, the Job is not finished.