API Reference

EasyJobsBase.Job — Type

Job(core::Thunk; description="", username="")

Create a simple job.

Arguments

core: a Thunk that encloses the job core definition.

name: give a short name to the job.

description::String="": describe what the job does in more detail.

username::String="": indicate who executes the job.

Examples

julia> using Thinkers

julia> a = Job(Thunk(sleep, 5); username="me", description="Sleep for 5 seconds");

julia> b = Job(Thunk(run, `pwd` & `ls`); username="me", description="Run some commands");

EasyJobsBase.:← — Method

←(y, x)

Chain two AbstractJobs reversely.

EasyJobsBase.:→ — Method

→(x, y)

Chain two AbstractJobs.

EasyJobsBase.chain! — Method

chain!(x::AbstractJob, y::AbstractJob, z::AbstractJob...)

Chain multiple AbstractJobs one after another.

EasyJobsBase.countexecution — Method

countexecution(job::AbstractJob)

Count how many times the job has been run.

EasyJobsBase.creationtimeof — Method

creationtimeof(job::AbstractJob)

Return the creation time of the job.

EasyJobsBase.descriptionof — Method

descriptionof(job::AbstractJob)

Return the description of the job.

EasyJobsBase.endtimeof — Method

endtimeof(job::AbstractJob)

Return the end time of the job. Return nothing if it has not exited.

EasyJobsBase.getstatus — Method

getstatus(job::AbstractJob)

Get the current status of the job.

EasyJobsBase.isexited — Method

isexited(job::AbstractJob)

Test if the job has exited.

EasyJobsBase.isfailed — Method

isfailed(job::AbstractJob)

Test if the job failed during running.

EasyJobsBase.isinterrupted — Method

isinterrupted(job::AbstractJob)

Test if the job was interrupted during running.

EasyJobsBase.ispending — Method

ispending(job::AbstractJob)

Test if the job is still pending.

EasyJobsBase.isrunning — Method

isrunning(job::AbstractJob)

Test if the job is running.

EasyJobsBase.issucceeded — Method

issucceeded(job::AbstractJob)

Test if the job was successfully run.

EasyJobsBase.kill! — Method

kill!(exec::Executor)

Manually kill a Job, works only if it is running.

EasyJobsBase.listexited — Method

listexited(jobs)

Filter the exited jobs in a sequence of jobs.

EasyJobsBase.listfailed — Method

listfailed(jobs)

Filter the failed jobs in a sequence of jobs.

EasyJobsBase.listinterrupted — Method

listinterrupted(jobs)

Filter the interrupted jobs in a sequence of jobs.

EasyJobsBase.listpending — Method

listpending(jobs)

Filter the pending jobs in a sequence of jobs.

EasyJobsBase.listrunning — Method

listrunning(jobs)

Filter the running jobs in a sequence of jobs.

EasyJobsBase.listsucceeded — Method

listsucceeded(jobs)

Filter the succeeded jobs in a sequence of jobs.

EasyJobsBase.run! — Method

run!(job::Job; maxattempts=1, interval=1, waitfor=0)

Run a Job with a maximum number of attempts, with each attempt separated by a few seconds.

EasyJobsBase.starttimeof — Method

starttimeof(job::AbstractJob)

Return the start time of the job. Return nothing if it is still pending.

EasyJobsBase.timecostof — Method

timecostof(job::AbstractJob)

Return the time cost of the job since it started running.

If nothing, the job is still pending. If it is finished, return how long it took to complete.

Thinkers.getresult — Method

getresult(job::AbstractJob)

Get the running result of the job.

The result is wrapped by a Some type. Use something to retrieve its value. If it is nothing, the job is not finished.